Refactor Qwen3 14B prefill fwd
Convert the Qwen3 14B prefill path to the jit/jit.inline full-layer structure and replace qwen3_14b_prefill.py with prefill_fwd.py.

Keep layer intermediate tensors inside the JIT path, produce final logits through RMSNorm plus LM head, and gather the final prefill token through a static [BATCH, HIDDEN] tensor before calling the shared rms_lm_head helper.

Address review feedback by deriving the block-table stride from the runtime block_table shape, removing the redundant input-hidden copy, and guarding zero-length sequences before computing the final token index.
